% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp22091R.R
\name{delta_i}
\alias{delta_i}
\title{Calculate Delta using R}
\usage{
delta_i(data_i, big_delta_symbolic, params_first, params_second, param_vals)
}
\arguments{
\item{data_i}{The i^th observation (Y_i, X_i, Z_i)}

\item{big_delta_symbolic}{Symbolic expression for Delta}

\item{params_first}{names of first stage parameters (theta) in a list}

\item{params_second}{names of second stage parameters (beta) in a list}

\item{param_vals}{estimated values of first and second stage parameters}
}
\value{
a matrix named big_delta of size row length(params_second) col length(params_first))
}
\description{
Calculate Delta from sample split paper for a single observation
}
