\name{StatComp22091-package}
\alias{StatComp22091-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  the function delta_i is one step in the estimate of a model,and _gibbs_cpp_ is a rcpp version to generate model with gibbs
  }
\author{
22091

Maintainer: 22091 <lxl2333@mail.ustc.edu.cn>
}
\keyword{ package }
